+++ Sources/PrimerSDK/Classes/Services/Network/PrimerAPIClient.swift
@@ -360,18 +360,6 @@ internal class PrimerAPIClient: PrimerAPIClientProtocol {
             }
         }
     }
-
-    func tokenizePaymentMethod(clientToken: DecodedClientToken, tokenizationRequestBody: Request.Body.Tokenization, completion: @escaping (_ result: Result<PrimerPaymentMethodTokenData, Error>) -> Void) {
-        let endpoint = PrimerAPI.tokenizePaymentMethod(clientToken: clientToken, tokenizationRequestBody: tokenizationRequestBody)
-        networkService.request(endpoint) { (result: Result<PrimerPaymentMethodTokenData, Error>) in
-            switch result {
-            case .success(let paymentMethodToken):
-                completion(.success(paymentMethodToken))
-            case .failure(let err):
-                completion(.failure(err))
-            }
-        }
-    }
     
     func createApayaSession(
         clientToken: DecodedClientToken,
